feat: 데일리 스크럼에서 오늘 할 일에는 상태가 DONE인 task를 추가할 수 없도록 수정 (CM-82)
